Corrigido erro no cáculo do tamanho da matriz. Adicionado debug_graph a makefile. Corrigido o cálculo da percentagem de arestas atribuídas.

+++ src/graph/graph.h
@@ -16,6 +16,7 @@
 #include <stdio.h>
 #include <math.h>
 #include <time.h>
+#include <stdint.h>
 
 #define MAX_EDGE 3
 #define MIN_EDGE 1
@@ -32,11 +33,11 @@ float *create_graph(int graph_size, int edge_percentage);
 float *create_locked_graph(int graph_size, int edge_percentage);
 void add_edge(float *graph, int u, int v, float weight);
 void add_null_edge(float *graph, int u, int v);
-unsigned long int get_matrix_size(int graph_size);
+unsigned int get_matrix_size(int graph_size);
 
 float get_edge(float *graph, int u, int v);
 int get_edge_count(float *graph, int graph_size);
-float get_edge_percentage(float *graph, int graph_size);
+double get_edge_percentage(float *graph, int graph_size);
 
 int pick_graph_root(int graph_size);
